Create STUDENT INFORMATION MANAGEMENT SYSTEM
This program is designed to manage information about a group of students using a structure in C.The program allows for the input of student information such as name, age, registration no, and course of study etc. It also includes a password protection feature to ensure that only authorized users can access the information. Once the user enters the correct password, they can search for a student by name and the program will return their information. This program can be used by teachers, administrators, or anyone who needs to keep track of student information.